+++ router/oauth.go
@@ -3,9 +3,7 @@ package router
 import (
 	"crypto/sha256"
 	"encoding/base64"
-	"errors"
 	"fmt"
-	"math/rand"
 	"net/http"
 	"os"
 
@@ -65,14 +63,4 @@ func (o *Oauth) GetCode(c echo.Context) error {
 	}
 
 	return c.String(http.StatusOK, authURL)
-}
-
-func RandomString(n int) string {
-	var letter = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
-
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = letter[rand.Intn(len(letter))]
-	}
-	return string(b)
 }
